requirements.txt
docopt==0.6.2
requests_html==0.10.0
requests==2.21.0
PyYAML==5.1.1
